refactor : 엔티티 타입변경에 따른 dto 타입변경

+++ backend/team6/src/main/java/programmers/team6/domain/vacation/rule/AnnualVacationRule.java
@@ -44,7 +44,7 @@ public final class AnnualVacationRule {
 	public VacationGrantInfo toGrantInfo(LocalDate now, VacationGrantEligibility vacationGrantEligibility) {
 		int yearsOfService = calcYearsOfService(now, vacationGrantEligibility.joinDateAsLocalDate());
 		return new VacationGrantInfo(vacationGrantEligibility.id(),
-			vacationGrantEligibility.vacationCount() + calcIncreaseDays(yearsOfService),
+			(int)vacationGrantEligibility.vacationCount() + calcIncreaseDays(yearsOfService),
 			vacationGrantEligibility.version());
 	}
