+++ backend/team6/src/main/java/programmers/team6/domain/vacation/rule/vacationgranteligiblities/VacationGrantInfo.java
@@ -1,4 +1,4 @@
 package programmers.team6.domain.vacation.rule.vacationgranteligiblities;
 
-public record VacationGrantInfo(Integer vacationInfoId, Integer totalCount, Integer version) {
+public record VacationGrantInfo(Integer vacationInfoId, double totalCount, Integer version) {
 }
